Amenagement de tous nos codes: ajout de mes trucs et ajustement de chacun de nos codes pour le bon fonctionnement

+++ src/main/java/universite_paris8/iut/yponnou/zelda/modele/Armes/Arme.java
@@ -2,10 +2,10 @@ package universite_paris8.iut.yponnou.zelda.modele.Armes;
 
 public class Arme{
     private String nom;
-    private double ptsDegats;
+    private int ptsDegats;
     private int portee;
 
-    public Arme(String nom, double ptsDegats, int portee) {
+    public Arme(String nom, int ptsDegats, int portee) {
         this.nom = nom;
         this.ptsDegats = ptsDegats;
         this.portee = portee;
@@ -14,7 +14,7 @@ public class Arme{
     public String getNom() {
         return nom;
     }
-    public double getPtsDegats() {
+    public int getPtsDegats() {
         return ptsDegats;
     }
     public int getPortee() {
