+++ src/main/java/universite_paris8/iut/yponnou/zelda/modele/Armes/ArmeDistance.java
@@ -3,7 +3,7 @@ package universite_paris8.iut.yponnou.zelda.modele.Armes;
 public class ArmeDistance extends Arme{
 
 
-    public ArmeDistance(String nom, double ptsDegats, int portee) {
+    public ArmeDistance(String nom, int ptsDegats, int portee) {
         super(nom, ptsDegats, portee);
     }
 }
